requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-allauth>=0.61
dj-rest-auth>=5.0
psycopg2-binary>=2.9
requests>=2.31
python-decouple>=3.8
